# services/api/openapi.json
{
  "components": {
    "schemas": {
      "AdminCaseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminCaseResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "AdminCaseListResponse",
        "type": "object"
      },
      "AdminCaseResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "seller_display_name": {
            "title": "Seller Display Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "type": {
            "$ref": "#/components/schemas/CaseType"
          }
        },
        "required": [
          "id",
          "order_id",
          "type",
          "status",
          "description",
          "created_at",
          "seller_display_name"
        ],
        "title": "AdminCaseResponse",
        "type": "object"
      },
      "AdminCaseUpdateBody": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "AdminCaseUpdateBody",
        "type": "object"
      },
...
      "CaseStatus": {
        "enum": [
          "open",
          "in_progress",
          "resolved"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
      "CaseType": {
        "enum": [
          "refund_request",
          "escalation",
          "other"
        ],
        "title": "CaseType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderItemResponse": {
        "properties": {
          "listing_id": {
            "format": "uuid",
            "title": "Listing Id",
            "type": "string"
          },
          "listing_sku": {
            "title": "Listing Sku",
            "type": "string"
          },
          "listing_title": {
            "title": "Listing Title",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "unit_price_cents": {
            "title": "Unit Price Cents",
            "type": "integer"
          }
        },
        "required": [
          "listing_id",
          "listing_title",
          "listing_sku",
          "quantity",
          "unit_price_cents"
        ],
        "title": "OrderItemResponse",
        "type": "object"
      },
      "OrderListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "OrderListResponse",
        "type": "object"
      },
      "OrderResponse": {
        "properties": {
          "buyer_ref": {
            "title": "Buyer Ref",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "placed_at": {
            "format": "date-time",
            "title": "Placed At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "total_cents": {
            "title": "Total Cents",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "buyer_ref",
          "status",
          "total_cents",
          "currency",
          "placed_at",
          "delivered_at",
          "updated_at",
          "items"
        ],
        "title": "OrderResponse",
        "type": "object"
      },
      "OrderStatus": {
        "enum": [
          "pending",
          "shipped",
          "delivered",
          "cancelled",
          "refund_requested",
          "refunded"
        ],
        "title": "OrderStatus",
        "type": "string"
      },
      "RefundRequestBody": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RefundRequestBody",
        "type": "object"
      },
      "SupportCaseResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "type": {
            "$ref": "#/components/schemas/CaseType"
          }
        },
        "required": [
          "id",
          "order_id",
          "type",
          "status",
          "description",
          "created_at"
        ],
        "title": "SupportCaseResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Seller Support Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/cases": {
      "get": {
        "operationId": "admin_list_cases_v1_admin_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCaseListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin List Cases",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/cases/{case_id}": {
      "patch": {
        "operationId": "admin_update_case_v1_admin_cases__case_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCaseUpdateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Update Case",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/cases/{case_id}": {
      "get": {
        "operationId": "get_case_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Case",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/v1/orders": {
      "get": {
        "operationId": "list_orders_v1_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrderStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Orders",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/{order_id}": {
      "get": {
        "operationId": "get_order_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/{order_id}/refund-requests": {
      "post": {
        "operationId": "create_refund_request_v1_orders__order_id__refund_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Refund Request",
        "tags": [
          "orders"
        ]
      }
    }
  }
}